requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.5.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
pyjwt==2.9.0
google-auth==2.33.0
anyio==4.4.0
pymongo==4.8.0
